🚚 rename: 파스칼케이스에서 케밥케이스로 파일 네이밍 규칙 변경
Cover.js -> cover.js
